ref: extract isConditionFulfilled util fn out
